refactor the cancellation token
